// src/store/user/user.types.ts
enum USERS_TYPES {
  LOGIN = 'USER/LOGIN',
  REGISTRATION = 'USER/REGISTRATION',
  SET_CREDENTIALS = 'USER/CREDENTIALS',
  SET_REGISTRATION_CREDENTIALS = 'USER/SET_REGISTRATION_CREDENTIALS',
  SET_LOADING = 'USER/LOADING',
  SET_USER_SUCCESS = 'USER/SUCCESS',
  SET_USER_FAILURE = 'USER/FAILURE',
  CLEAR_STATE = 'USER/CLEAR',
  SET_USER_ID = 'USER/SET_ID',
}

export default USERS_TYPES;